refactor: View설계 단순화하기
