+++ src/main/java/view/inputview/InputView.java
@@ -1,5 +1,9 @@
 package view.inputview;
 
-public interface InputView<T> {
-    T getUserInputData();
+import java.util.Set;
+
+public interface InputView {
+    int inputPurchaseMoney();
+    int inputBonusNumber();
+    Set<Integer> inputWinningNumbers();
 }
