+++ src/main/java/view/outputview/OutputView.java
@@ -1,5 +1,11 @@
 package view.outputview;
 
-public interface OutputView<P> {
-    void showOutputData(P parameter);
+import java.util.List;
+import model.LottoTicketDto;
+import model.WinningResultDto;
+
+public interface OutputView {
+    void showLottoTicket(List<LottoTicketDto> lottoTickets);
+    void showWinningResult(List<WinningResultDto> winningResults);
+    void showRateOfReturn(double rateOfReturn);
 }
